refactor (network): move part network logic to repository
