+++ personalization-sdk/src/main/kotlin/com/personalization/sdk/data/di/RepositoriesModule.kt
@@ -1,8 +1,10 @@
 package com.personalization.sdk.data.di
 
+import com.personalization.sdk.data.repositories.network.NetworkRepositoryImpl
 import com.personalization.sdk.data.repositories.notification.NotificationRepositoryImpl
 import com.personalization.sdk.data.repositories.preferences.PreferencesRepositoryImpl
 import com.personalization.sdk.data.repositories.recommendation.RecommendationRepositoryImpl
+import com.personalization.sdk.domain.repositories.NetworkRepository
 import com.personalization.sdk.domain.repositories.NotificationRepository
 import com.personalization.sdk.domain.repositories.PreferencesRepository
 import com.personalization.sdk.domain.repositories.RecommendationRepository
@@ -14,6 +16,9 @@ abstract class RepositoriesModule {
 
     @Binds
     abstract fun bindPreferencesRepository(impl: PreferencesRepositoryImpl): PreferencesRepository
+
+    @Binds
+    abstract fun bindNetworkRepository(impl: NetworkRepositoryImpl): NetworkRepository
 
     @Binds
     abstract fun bindRecommendationRepository(impl: RecommendationRepositoryImpl): RecommendationRepository
