+++ personalization-sdk/src/main/kotlin/com/personalization/sdk/domain/usecases/network/SendNetworkMethodUseCase.kt
@@ -1,0 +1,24 @@
+package com.personalization.sdk.domain.usecases.network
+
+import com.personalization.api.OnApiCallbackListener
+import com.personalization.sdk.domain.models.NetworkMethod
+import com.personalization.sdk.domain.repositories.NetworkRepository
+import org.json.JSONObject
+import javax.inject.Inject
+
+class SendNetworkMethodUseCase @Inject constructor(
+    private val networkRepository: NetworkRepository
+) {
+
+    fun invoke(
+        networkMethod: NetworkMethod,
+        params: JSONObject,
+        listener: OnApiCallbackListener?
+    ) {
+        networkRepository.sendMethod(
+            networkMethod = networkMethod,
+            params = params,
+            listener = listener
+        )
+    }
+}
